backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pyyaml==6.0.2
google-genai==0.4.0
pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
